backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.0
aiomysql==0.2.0
redis==5.2.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
slowapi==0.1.9
argon2-cffi==23.1.0
python-magic==0.4.27
pyotp==2.9.0
pip-audit==2.7.3
psutil==5.9.8
arq==0.26.1
Pillow==10.2.0
brotli-asgi==1.4.0
aioboto3==12.3.0
python-jose[cryptography]
python-multipart
email-validator==2.2.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
aiosqlite==0.20.0
pytest-mock==3.12.0
docker==7.1.0
GitPython==3.1.43